rename load_type -> load, tangential method= kwarg
also version bump

Why split get from get_path? The program knows what it needs, and that
is one way of forcing a bit of order and expectation into the config.
Still not sure if it is the right design decision, especially since it
makes it hard to communicate about that is expected in one place,
instead it is worst case scattered all over the place, or best case
centralized at decorator definition time ... maybe we want it in both
places?