fix get_path behavior for defaults from user configs, ver bump
Default paths defined in user configs were being expanded using the
auth config pathit function. This is bad.

The fix was to handle the for_path branch explicitly in AuthConfig.get
(where the source of a variable is still known) when adding values to
the list of defaults.

By calling the source pathit the fully expanded path will take
priority when the second call to pathit in get_path tries to expand
the path a second time because the the path starts with a slash. We
should not be silently relying on this behavior.

The real way to fix this is the version of this that retains source
information until the very end.

also version bump to get this fix out