actually fix get_path behavior for defaults, version bump
As it turns out #3 was caused by yet another manifestation of the
issue that was encountered in a8ef37a -- the incorrectly handled
branches inside of get. The internal implementation is pretty much
entirely inside out, the version where we retain the sources, rank by
source and then pick the resulting value will end up being much less
code, the inner block that this commit touches how has the same
pattern repeated something like 8 times varying across string vs list
and path vs not path, and probably one other variable, but all the
logic is the same, it was just hard to track.

Closes #3.