make sure we never write an empty user config
While investigating https://github.com/tgbugs/orthauth/issues/7 I came
across an instance where a file was opened for writing in a context
manager and there was a call inside the context manager that could
fail, resulting in an empty file being written. Remind me again why
python allows that in the first place? Oh right, streaming writes and
the impossiblity of knowing whether an error might be raised by even
something seemingly innocuous like json.dump. No truncated files here,
just empty files! This is something for a linter, waren on  function
calls made inside a context manager that opens a file for writing.

In the process I also discovered that the DANGERZONE on the pyyaml
import is no joke. yaml.safe_load('') -> None. Thus have added a check
to load so that if None is ever returned by one of the format readers
it will be caught before it can cause problems downstream. Also don't
set self._blob until that check passes in the event that some hapless
soul catches the load error and would then be stuck with a bad value
for self._blob.