tests for empty config, call self.load() in *Config.__new__
The easiest way to make check all the configs is to call self.load()
at the end of new if self._path is not None. Probably need to do a
performance check on this weighed against using auth.audit. The if the
performance cost is too much then we will need to uncomment the call
that would fail the first time a query goes to the user config.

Also version bump for release.

closes https://github.com/tgbugs/orthauth/issues/4